pyproject.toml
[project]
name = "snraware"
description = "A deep learning imaging AI framework"
version = "0.1.0"
readme = "README.md"
requires-python = ">=3.12,<4.0"
authors = [
    {name = "Hui Xue", email = "xueh@microsoft.com"},
]
keywords = []
license = "MIT"
dependencies = [
    "mkl-fft>=2.0.0",
    "mkl-service>=2.4.2",
    "typing_extensions",
    "click",
    "h5py",
    "opencv-python",
    "pathlib",
    "PyWavelets",
    "pytorch-wavelets",
    "torchmetrics==1.7.3",
    "torch-interpol",
    "piq",
    "imageio",
    "pyyaml",
    "wandb",
    "wandb[media]",
    "scipy",
    "cffi",
    "numpy>=2.2.6",
    "torch>=2.8.0",
    "setuptools",
    "torchvision",
    "torchinfo",
    "prettytable",
    "scikit-image",
    "nibabel",
    "matplotlib",
    "einops",
    "pandas",
    "lightning",
    "hydra-core",
    "omegaconf",
    "hydra_colorlog",
    "pytest>=8.4.2",
    "dvc[azure]",
]

[project.urls]
repository = "https://github.com/microsoft/SNRAware"
documentation = "https://microsoft.github.io/SNRAware"

[dependency-groups]
dev = [
    "coverage",
    "pyright",
    "pytest",
    "pytest-cov",
    "pytest-nunit",
    "setuptools",
    "ruff",
    # documentation tooling
    "mkdocs",
    "mkdocs-awesome-pages-plugin",
    "mkdocs-material",
    "mkdocs-video",
    "mkdocs-glightbox",
    "mdx-truly-sane-lists",
    "mkdocstrings",
    "mkdocstrings-python",
]

[build-system]
requires = ["uv_build>=0.9.2,<0.10.0"]
build-backend = "uv_build"

[tool.uv.build-backend]
module-name = ["snraware"]

[[tool.uv.index]]
name = "mkl"
url = "https://urob.github.io/numpy-mkl"

[tool.uv.sources]
numpy = { index = "mkl" }
scipy = { index = "mkl" }
mkl-service = { index = "mkl" }

[project.scripts]

[tool.pyright]
#pythonVersion = "3.10"
include = ["src", "test"]
exclude = ["**/__pycache__"]

deprecateTypingAliases = true
reportPropertyTypeMismatch = true
reportUnusedImport = true
reportDuplicateImport = true
reportUntypedFunctionDecorator = true
reportUntypedClassDecorator = true
reportUntypedBaseClass = true
reportUntypedNamedTuple = true
reportDeprecated = true
reportInconsistentConstructor = true
reportUnnecessaryCast = true
reportUnnecessaryContains = true
reportImplicitStringConcatenation = true
reportInvalidStubStatement = true
reportIncompleteStub = true
reportUnnecessaryTypeIgnoreComment = true
reportMatchNotExhaustive = true

reportMissingTypeStubs = false
reportImportCycles = true
reportUnusedFunction = false
reportMissingSuperCall = false
reportUnknownLambdaType = true
reportCallInDefaultInitializer = true
reportUnnecessaryIsInstance = true
reportUnusedCallResult = false
reportUnusedExpression = true
reportImplicitOverride = false
reportIncompatibleVariableOverride = true
reportCallIssue = false
reportGeneralTypeIssues = false
reportIndexIssue = false
reportOptionalIterable = false
reportOperatorIssue = false
reportOptionalCall = false
reportMissingTypeArgument = false

reportIncompatibleMethodOverride = false
reportUninitializedInstanceVariable = false
reportMissingParameterType = false
reportUnknownArgumentType = false
reportUnknownMemberType = false
reportConstantRedefinition = false
reportUnknownVariableType = false
reportUnknownParameterType = false
reportArgumentType = false
reportUnusedVariable = false
reportPossiblyUnboundVariable = false
reportAttributeAccessIssue = false
reportOptionalMemberAccess = false
reportPrivateUsage = false
reportOptionalSubscript = false
reportUnusedClass = false
reportUndefinedVariable = false
reportUnnecessaryComparison = false

[tool.ruff]
cache-dir = ".tmp/.ruff_cache"
include = [
    "src/**/*.py",
    "test/**/*.py",
    "pyproject.toml",
]
fix = true
show-fixes = true
output-format = "full"
line-length = 99
indent-width = 4
target-version = "py312"

[tool.ruff.lint]
select = [
    "B",    # Bugbear
    "D",    # Docstring
    "E",    # Pycodestyle
    "F",    # Pyflakes
    "I",    # Import
    "Q",    # Quotes
    "W",    # PEP8 warnings
    "NPY",  # Numpy
    "PD",   # Pydocstyle
    "PTH",  # Path
    "RUF",  # Ruff
    "UP",   # Unittest
]
ignore = [
    "E501", # E501: Line too long ({width} > {limit})
    "D212", # D212: Multi-line docstring summary should start at the second line
    # This rule seems buggy: https://github.com/astral-sh/ruff/issues/2480
    "PD011", # PD011: Use `.to_numpy()` instead of `.values`
    # TODO: re-enable these when we can fix the errors
    "B007",  # Loop control variable `loader_ind` not used within loop body
    "B008",  # Do not perform function call `torch.device` in argument defaults
    "B016",  # Cannot raise a literal. Did you intend to return it or raise an Exception?
    "D100",  # Missing docstrings
    "D101",  # Missing docstrings
    "D102",  # Missing docstrings
    "D103",  # Missing docstrings
    "D104",  # Missing docstrings
    "D105",  # Missing docstrings
    "D107",  # Missing docstrings
    "D205",  # 1 blank line required between summary line and description
    "D417",  # Missing argument descriptions in the docstring
    "E722",  # Do not use bare `except`
    "F403",  # Do not use `*` imports
    "F405",  # X may be undefined, or defined from star imports
    "PTH110", # `os.path.exists()` should be replaced by `Path.exists()`
    "PTH112", # `os.path.isdir()` should be replaced by `Path.is_dir()`
    "PTH113", # `os.path.isfile()` should be replaced by `Path.is_file()`
    "PTH118", # `os.path.join()` should be replaced by `Path` with `/` operator
    "PTH103", # `os.makedirs()` should be replaced by `Path.mkdir(parents=True)`
    "PTH123", # `open()` should be replaced by `Path.open()`
    "PTH208", # Use `pathlib.Path.iterdir()` instead.
    "PTH120", # `os.path.dirname()` should be replaced by `Path.parent`
    "NPY002",
]

# Allow fix for all enabled rules (when `--fix`) is provided.
fixable = ["ALL"]
unfixable = []

# Allow unused variables when underscore-prefixed.
dummy-variable-rgx = "^(_+|(_+[a-zA-Z0-9_]*[a-zA-Z0-9]+?))$"

[tool.ruff.lint.per-file-ignores]
"test/*" = [
    "D100", # Missing docstring in public module
    "D101", # Missing docstring in public class
    "D102", # Missing docstring in public method
    "D103", # Missing docstring in public function
    "D104", # Missing docstring in public package
    "E501", # Line too long,
    "D205", # blank line required between summary line and description
    "D107", # Missing docstring in `__init__`
]

[tool.ruff.lint.pydocstyle]
convention = "google"

[tool.ruff.lint.isort]
known-first-party = ["snraware", "snraware_test"]

[tool.ruff.format]
quote-style = "double"            # Like Black, use double quotes for strings.
indent-style = "space"            # Like Black, indent with spaces, rather than tabs.
skip-magic-trailing-comma = false # Like Black, respect magic trailing commas.
line-ending = "auto"              # Like Black, automatically detect the appropriate line ending.

# Enable auto-formatting of code examples in docstrings. Markdown,
# reStructuredText code/literal blocks and doctests are all supported.
docstring-code-format = true

[tool.coverage.run]
data_file = ".tmp/.coverage"
relative_files = true

[tool.coverage.report]
show_missing = true

[tool.coverage.xml]
output = ".tmp/reports/coverage.xml"

[tool.pytest.ini_options]
cache_dir = ".tmp/.pytest_cache"
testpaths = ["test"]
addopts = """
    --cov=src/snraware --cov=test/snraware_test
    --cov-report=term
    --cov-report=xml
    --junit-xml .tmp/reports/pytest-junit.xml
    --capture=no
    --cov-branch
    --cache-clear
    -m 'not gpu'
    -s
"""
markers = [
    "gpu: a test needs gpus to run",
    "slow: mark a test as slow",
]
